requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
playwright
python-multipart
httpx
google-auth-oauthlib
google-api-python-client
